+++ oauthproxy.go
@@ -18,7 +18,6 @@ import (
 	"time"
 
 	"github.com/coreos/go-oidc"
-
 	"github.com/mbland/hmacauth"
 	sessionsapi "github.com/pusher/oauth2_proxy/pkg/apis/sessions"
 	"github.com/pusher/oauth2_proxy/pkg/cookies"